+++ src/components/20-molecules/datepicker/test.accessor.js
@@ -97,10 +97,10 @@ class DatePickerAccessor {
     if (background === 'rgba(0, 0, 0, 0)' && color === 'rgb(51, 51, 51)') {
       return 'STANDARD';
     } else if (
-      background === 'rgb(204, 204, 204)' &&
+      background === 'rgb(229, 229, 229)' &&
       color === 'rgb(51, 51, 51)'
     ) {
-      return 'PRESELECTED';
+      return 'TODAY';
     } else if (
       background === 'rgb(0, 0, 143)' &&
       color === 'rgb(255, 255, 255)'